// client/src/pages/Crm/Vendors/types.ts
export interface BenefitsFormState {
  eqpPricing: number;
  rebatePercentage: number;
  freeSetups: boolean;
  freeSpecSamples: boolean;
  reducedSpecSamples: boolean;
  freeSelfPromo: boolean;
  reducedSelfPromo: boolean;
  ytdEqpSavings: number;
  ytdRebates: number;
  selfPromosSent: number;
  specSamplesSent: number;
}